requirements.txt
Flask==3.0.0
grpcio==1.59.0
grpcio-tools==1.59.0
protobuf==4.24.4
nlohmann-json==0.0.1
requests==2.31.0
python-dateutil==2.8.2